+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * builder/Azure: add support for packer's azure builder
 * builder/CloudStack: add support for packer's cloudstack builder
 * builder/DigitalOcean: add support for packer's digitalocean builder
+* builder/Null: add support for packer's null builder
 * builder/ProfitBricks: add support for packer's ProfitBricks builder
 * builder/Qemu: add support for packer's qemu builder
 * builder/Triton: add support for packer's triton builder